Change distanceA2AEuclideanSquared method to store only needed values
distanceA2AEuclideanSquared method created the entire upper triangular
matrix and computed all the distances between all the vectors. But the
method response was only a collection of distances from the vector i to
the vector i+1.
The method no longer computes all the distances but only the distances
needed for the output.