Fix checkFeasibility method on AreaManager
The method was failing when a region was composed by two islands
of more than 1 areas. In that case what happended was that each
island's area was adding it's neighbors to the complete set of
regions, so the difference between the supposed regions and the
neighboring regions was always empty.
The fix came with a stack of areas when only neighbors of the first
visited region got added. After visiting and adding all the areas
adjacent to the first area, a difference of the supposed areas and
the visited areas should give empty to be valid.