Modifications in the neighbors function
Changed some names of the variables and started to remove the third
unnecessary parameter in the checkFeasibility function.

Changed also the way the result was being generated, now a dictionary
if filled and then given to the class. Before, the instance was being
updated every time a new result appeared.