Set number of cores to size of processes pool
Also changed the way that the random received the seed.
Before the problem was that when using the same 'session' or
script, the process ID was used as seed, giving the same result for the
randoms.

Added test to ensure that we get different results from the same
execution.
This should fix #17
Need to confirm that also is good for #16