Add parameters to viz functions
Now I can specify the values for the colors, the color map (from the
matplotlib's colormaps) and the alpha for the colors.

The values for the colors is to give each area an 'intensity'. This
is useful to colorize a variable, so that the intensity of the color
represents the value of the variable.